backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
PyPDF2
pdfplumber
sentence-transformers
faiss-cpu
openai
tiktoken
numpy
pydantic
